JPA Repository. JPQL, QueryDSL 중 선택하여 DB 조회하는 Service 작성

+++ src/main/java/com/example/arnoldkimcommunitybe/practice/JPAPracticeRepository.java
@@ -2,6 +2,6 @@ package com.example.arnoldkimcommunitybe.practice;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JPAPracticeRepository extends JpaRepository<PracticeEntity, Long> {
+public interface JPAPracticeRepository extends JpaRepository<PracticeEntity, Integer> {
 
 }
